bot/requirements.txt
python-dotenv
psycopg2
python-telegram-bot



# async def send_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
#     message_text = (
#         "Get Edified🥰🙏 !!!\n\n\n"
#         "We’ll be going before the Lord by 5:30pm today with hearts ready to seek Him. "
#         "Tell a Prayer Force member to tell a Prayer Force member to tell another Prayer Force member "
#         "that today's meeting is not one you want to miss."
#     )

#     await context.bot.send_message(
#         chat_id=update.effective_chat.id,
#         text=message_text

#     )
#     await context.bot.send



# import asyncio
# import telegram
# from telegram import Update

# async def main():
#     bot = telegram.Bot("7558528054:AAGZhgpDKKVI3TZdsg3iEVzSh-wtcaf707Q")
#     async with bot:
#         print(await bot.get_me())
#         chat_id = Update.effective_chat.id
#         await bot.send_message(chat_id, text='Hello, world!')


# async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
#     await context.bot.send_message(
#         chat_id=update.effective_chat.id,
#         text="I'm a bot, please talk to me!"
#     )

# if __name__ == '__main__':
#     asyncio.run(main())


# import asyncio